Prefer timezone converter to fallback on config default
Previously 'mon' may be treated as a source timezone if used as part of conversion string 'mon 7pm est' as it parses the line as 'est' the destination zone, 7pm is the 'joiner' word, and 'mon' was the source zone. 'Mon' matches multiple zone candidates so never entered the peferred default converter.